+++ server/src/main/java/org/allaymc/server/component/annotation/Manager.java
@@ -1,6 +1,6 @@
 package org.allaymc.server.component.annotation;
 
-import org.allaymc.api.component.interfaces.ComponentManager;
+import org.allaymc.server.component.interfaces.ComponentManager;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;