+++ server/src/main/java/org/allaymc/server/entity/component/EntityAttributeComponentImpl.java
@@ -1,7 +1,6 @@
 package org.allaymc.server.entity.component;
 
 import lombok.extern.slf4j.Slf4j;
-import org.allaymc.api.component.interfaces.ComponentManager;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.component.attribute.Attribute;
 import org.allaymc.api.entity.component.attribute.AttributeType;
@@ -11,6 +10,7 @@ import org.allaymc.api.eventbus.event.entity.EntityHealthChangeEvent;
 import org.allaymc.api.utils.Identifier;
 import org.allaymc.server.component.annotation.ComponentObject;
 import org.allaymc.server.component.annotation.Manager;
+import org.allaymc.server.component.interfaces.ComponentManager;
 import org.allaymc.server.entity.component.event.CEntityAttributeChangeEvent;
 import org.allaymc.server.entity.component.event.CEntityLoadNBTEvent;
 import org.allaymc.server.entity.component.event.CEntitySaveNBTEvent;