+++ server/src/main/java/org/allaymc/server/network/processor/impl/login/LoginPacketProcessor.java
@@ -76,18 +76,17 @@ public class LoginPacketProcessor extends ILoginPacketProcessor<LoginPacket> {
         }
 
         try {
+            networkComponent.setNetworkEncryptionEnabled(true);
+
             var clientKey = EncryptionUtils.parseKey(loginData.getIdentityPublicKey());
             var serverKeyPair = EncryptionUtils.createKeyPair();
             var token = EncryptionUtils.generateRandomToken();
-            var encryptionSecretKey = EncryptionUtils.getSecretKey(serverKeyPair.getPrivate(), clientKey, token);
-
-            networkComponent.setEncryptionSecretKey(encryptionSecretKey);
-            networkComponent.setNetworkEncryptionEnabled(true);
 
             var handshakePacket = new ServerToClientHandshakePacket();
             handshakePacket.setJwt(EncryptionUtils.createHandshakeJwt(serverKeyPair, token));
             player.sendPacketImmediately(handshakePacket);
 
+            var encryptionSecretKey = EncryptionUtils.getSecretKey(serverKeyPair.getPrivate(), clientKey, token);
             player.getClientSession().enableEncryption(encryptionSecretKey);
             // completeLogin() when client send back ClientToServerHandshakePacket
         } catch (Exception exception) {